% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAnnualDegreeDays.R
\name{aggCells}
\alias{aggCells}
\title{Aggregate cellular HDD/CDD values to country-wide average (population-weighted)}
\usage{
aggCells(data, weight, mask)
}
\arguments{
\item{data}{SpatRaster object containing HDD/CDD values}

\item{weight}{SpatRaster object containing aggregation weights}

\item{mask}{SpatRaster object defining (regional) aggregation boundaries}
}
\value{
data frame containing regionally averaged HDD/CDD values
}
\description{
Aggregate cellular HDD/CDD values to country-wide average (population-weighted)
}
\author{
Hagen Tockhorn
}
