% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeBAITpars.R
\name{computeBAITpars}
\alias{computeBAITpars}
\title{Calculate Regression Parameters for BAIT Climate Variables}
\usage{
computeBAITpars(model = "20crv3-era5", cacheDir = "inst/intdata/BAITpars")
}
\arguments{
\item{model}{Character string specifying the GCM responsible for data input.}

\item{cacheDir}{Character string specifying the directory containing pre-calculated
function outputs.}
}
\value{
A `terra::SpatRaster` object with layers representing regression
parameters for each cell.
}
\description{
Performs linear regression on historical climate data to calculate
regression parameters for key climate variables: surface downwelling shortwave
radiation (rsds), near-surface wind speed (sfcwind), and near-surface specific
humidity (huss), with respect to near-surface air temperature (tas).

The regression utilizes a simple linear model based on historical data from
2000 to 2019-21 depending on the model. For rsds and sfcwind, a linear relationship
is assumed, while for huss, an exponential relationship is adopted, reflecting
the non-linear dependence of water vapor pressure on temperature.
}
\author{
Hagen Tockhorn
}
