% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmUtils.R
\name{createSlurm}
\alias{createSlurm}
\title{Create a Slurm Job for Running initCalculation}
\usage{
createSlurm(
  fileRow,
  pop,
  ssp,
  bait,
  tLim,
  hddcddFactor,
  wBAIT,
  jobConfig = list(),
  outDir = "output"
)
}
\arguments{
\item{fileRow}{A data frame row containing file mapping details}

\item{pop}{Population data to be passed to initCalculation}

\item{ssp}{Shared Socioeconomic Pathway scenario}

\item{bait}{Logical indicating whether to use BAIT calculation}

\item{tLim}{Temperature limits for degree day calculation}

\item{wBAIT}{Weights for BAIT calculation}

\item{jobConfig}{List of Slurm job configuration parameters. Currently supports}

\item{outDir}{A string specifying the absolute path to the \code{output} directory. This will also
contain the \code{logs} and \code{tmp} directories for log and temporary files.}

\item{hddCddFactor}{Heating/Cooling Degree Day factors}
}
\value{
A list containing job details including job name, script path,
        output file path, and Slurm submission command
}
\description{
Create a Slurm Job for Running initCalculation
}
\author{
Hagen Tockhorn
}
