% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDegreeDays.R
\name{getDegreeDays}
\alias{getDegreeDays}
\title{Calculate Heating and Cooling Degree Days}
\usage{
getDegreeDays(
  mappingFile = NULL,
  bait = TRUE,
  tLim = list(HDD = seq(9, 19), CDD = seq(15, 25)),
  std = c(tLim = 2, tAmb = 2),
  ssp = c("historical", "SSP2"),
  outDir = "output"
)
}
\arguments{
\item{mappingFile}{A string specifying the path to the mapping file containing information about the climate data
Must include columns like \code{"gcm"}, \code{"rcp"}, \code{"start"}, \code{"end"}, \code{"tas"}, \code{"rsds"},
\code{"sfcwind"} and \code{"huss"}.}

\item{bait}{Logical. If \code{TRUE}, bias-adjusted internal temperature (BAIT)
is included in the calculations. Defaults to \code{TRUE}.}

\item{tLim}{A list defining temperature limits for HDD and CDD calculations.
Defaults to \code{list("HDD" = seq(9, 19), "CDD" = seq(15, 25))}.}

\item{std}{A named vector of standardization parameters for temperature limits and ambient temperatures.
Defaults to \code{c("tLim" = 2, "tAmb" = 2)}.}

\item{ssp}{A character vector specifying the SSP scenarios to include. Defaults to \code{c("historical", "SSP2")}.}

\item{outDir}{A string specifying the absolute path to the \code{output} directory. This will also
contain the \code{logs} and \code{tmp} directories for log and temporary files.}
}
\description{
This function initiates the calculation of annual Heating (HDDs) and Cooling Degree Days (CDDs)
for historical and future scenarios. The calculation supports raw near-surface
air temperature and the bias-adjusted internal temperature (BAIT) which incorporates
additional climate variables.
The user can specify Shared Socioeconomic Pathways (SSPs) to define future population
development. The corresponding climate scenarios, represented by Representative Carbon Pathways (RCPs),
are chosen according to the IPCC scenario matrix.
Annual degree days are calculated per data set (typically 10-year period) in an
individually defined and submitted SLURM job. After successful completion, the job
outputs are gathered, post-processed and saved as a .csv file in the \code{output} folder.
}
\author{
Hagen Tockhorn
}
